feat: only the user owner can now delete his account

+++ src/controllers/userController.js
@@ -50,8 +50,16 @@ exports.updateUser = async (request, response) => {
 }
 
 exports.deleteUser = async (request, response) => {
+    const userId = request.params.id
+    const loggedInUserId = request.user._id
+
+    // On vérifie que l'utilisateur qui fait la requête est bien le propriétaire du compte
+    if (userId !== loggedInUserId.toString()) {
+        return response.status(403).json({ message: 'You can only delete your own account' })
+    }
+
     try {
-        const user = await User.findByIdAndDelete(request.params.id)
+        await User.findByIdAndDelete(request.params.id)
 
         response.status(200).json({ message: "User deleted successfully" })
     } catch (error) {
